[addbook_search] codesplit search from AddBookPage into QuickBookSearch and disable it.
